Error Responses On Bad Inputs

+++ back/nest-app-project/src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { UserEntity } from './user.entity'
@@ -19,16 +19,25 @@ export class UserService {
   }
 
   async findOne(id: string) {
-    return await this.repository.findOne({ where: { id } });
+    const user = await this.repository.findOne({ where: { id } });
+    if (!user)
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    return user;
   }
 
   async update(id: string, data: Partial<UserDTO>) {
+    const user = await this.repository.findOne({ where: { id } });
+    if (!user)
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
     await this.repository.update({ id }, data);
-    return await this.repository.findOne({ where: { id } });
+    return user;
   }
 
   async remove(id: string) {
+    const user = await this.repository.findOne({ where: { id } });
+    if (!user)
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
     await this.repository.delete({ id });
-    return { deleted: true };
+    return user;
   }
 }
